refactored some files for better organization
