+++ app/src/main/java/com/example/signifybasic/signrecognition/liverecognition/LiveSignRecognitionActivity.kt
@@ -1,4 +1,4 @@
-package com.example.signifybasic.signrecognition
+package com.example.signifybasic.signrecognition.liverecognition
 
 import android.Manifest
 import android.content.pm.PackageManager
