+++ app/src/main/java/com/example/signifybasic/signrecognition/videorecognition/ModelApiService.kt
@@ -1,4 +1,4 @@
-package com.example.signifybasic.signrecognition
+package com.example.signifybasic.signrecognition.videorecognition
 
 import okhttp3.MultipartBody
 import okhttp3.ResponseBody
