+++ app/src/main/res/layout/activity_record_video.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:padding="16dp"
     android:background="@drawable/login_background"
-    tools:context=".signrecognition.RecordVideoActivity">
+    tools:context=".signrecognition.videorecognition.RecordVideoActivity">
 
     <!-- loading animation -->
     <ProgressBar
