+++ app/src/main/res/layout/activity_sign_recognition_result.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:padding="16dp"
     android:background="@drawable/login_background"
-    tools:context=".signrecognition.SignRecognitionResultActivity">
+    tools:context=".signrecognition.videorecognition.SignRecognitionResultActivity">
 
     <!-- Recognized Sign Output -->
     <TextView
